requirements.txt
InquirerPy==0.3.4
Pillow==11.2.1
python_bcrypt==0.3.2
rich==14.0.0
